+++ database/migrations/2022_08_10_171336_add_input_discount_on_sale_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('sales', function (Blueprint $table) {
-            $table->integer('discount')->nullable()->default(0);
+            $table->decimal('discount', 10, 2)->nullable()->default(0);
         });
     }
