+++ database/migrations/2023_06_10_193111_drop_company_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
             $table->string('cnpj_cpj', 18)->unique()->nullable();
             $table->string('ie', 18)->unique()->nullable();
             $table->string('im', 18)->unique()->nullable();
-            $table->integer('balance')->nullable()->default(0);
+            $table->decimal('balance', 10, 3)->nullable()->default(0);
             $table->timestamp('validate')->nullable();
             $table->boolean('active')->nullable()->default(1);
         });
